Refactor: 이미지 Facade 수정
