+++ src/main/java/com/yeojeong/application/domain/board/board/application/imagefacade/ImageFacade.java
@@ -1,0 +1,7 @@
+package com.yeojeong.application.domain.board.board.application.imagefacade;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface ImageFacade {
+    String saveImage(MultipartFile image);
+}
